% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EW.R
\name{EW}
\alias{EW}
\alias{dEW}
\alias{pEW}
\alias{qEW}
\alias{rEW}
\title{The Exponentiated Weibull Distribution}
\usage{
dEW(x, alpha, theta, lambda, log = FALSE)

pEW(q, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

qEW(p, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

rEW(n, alpha, theta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{scale parameter.}

\item{theta,lambda}{shape parameters.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
Sabes una cosa...
}
\description{
Density, distribution function, quantile function and
random generation for the exponentiated weibull  distribution with
parameters \code{alpha}, \code{theta} and \code{lambda}.
}
\details{
The Exponentiated Weibull Distribution with parameters \code{alpha},
\code{theta} and \code{lambda} has density given by

f(x)=lambda*alpha*theta*x^(theta-1)*exp(-alpha*(x^theta))*(1-exp(-alpha*(x^theta)))^(lambda-1)
}
\examples{
## Curve
curve(dEW(x,2,1.5,0.5), from=0, to=2, col="red",ylab="Density")

## Curve
curve(pEW(x,alpha=2,theta=1.5,lambda=0.5), from=0, to=2,  col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qEW(p,alpha=2,theta=1.5,lambda=0.5), y=p)

## Curve
hist(rEW(10000,alpha=2,theta=1.5,lambda=0.5),freq=F,xlab="x", main="")
}

