% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FWE.R
\name{FWE}
\alias{FWE}
\alias{dFWE}
\alias{hFWE}
\alias{pFWE}
\alias{qFWE}
\alias{rFWE}
\title{The Flexible Weibull Extension Distribution}
\usage{
dFWE(x, alpha, beta, log = FALSE)

pFWE(q, alpha, beta, lower.tail = TRUE, log.p = FALSE)

qFWE(p, alpha, beta, lower.tail = TRUE, log.p = FALSE)

rFWE(n, alpha, beta)

hFWE(x, alpha, beta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{beta}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dFWE} gives the density, \code{pFWE} gives the distribution
function, \code{qFWE} gives the quantile function, \code{rFWE}
generates random deviates and \code{hFWE} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation  and hazard function for the flexible weibull extension distribution with
parameters \code{alpha} and  \code{beta}.
}
\details{
The flexible weibull extension with parameters \code{alpha} and \code{beta}
has density given by

f(x) = (alpha + (beta/x^2))*exp(alpha*x - beta/x)*exp(-exp(alpha*x-beta/x))

for x>0.
}
\examples{
## The probability density function
curve(dFWE(x, alpha = 0.75, beta = 0.5), from = 0, to = 3, ylim = c(0, 1.7), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pFWE(x, alpha = 0.75, beta = 0.5), from = 0, to = 3, col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pFWE(x, alpha = 0.75, beta = 0.5, lower.tail = FALSE), from = 0, to = 3, col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qFWE(p,alpha = 0.75, beta = 0.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pFWE(x, alpha = 0.75, beta = 0.5), from = 0, add = TRUE, col = "red")

## The random function
hist(rFWE(n = 1000, alpha = 2, beta = 0.5), freq = FALSE, xlab = "x", ylim = c(0, 2), las = 1, main = "")
curve(dFWE(x, alpha = 2, beta = 0.5),  from = 0, to = 3, add = TRUE, , col = "red")

## The Hazard function
curve(hFWE(x, alpha = 0.75, beta = 0.5), from = 0, to = 2, ylim = c(0, 2.5), col = "red", ylab = "The Hazard function", las = 1)
}

