% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMW.R
\name{GMW}
\alias{GMW}
\alias{dGMW}
\alias{hGMW}
\alias{pGMW}
\alias{qGMW}
\alias{rGMW}
\title{The Generalized modified Weibull Distribution}
\usage{
dGMW(y, beta, theta, gamma, lambda, log = FALSE)

pGMW(q, beta, theta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

qGMW(p, beta, theta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

rGMW(n, beta, theta, gamma, lambda)

hGMW(x, beta, theta, gamma, lambda, log = FALSE)
}
\arguments{
\item{beta}{parameter one.}

\item{theta}{parameter two.}

\item{gamma}{parameter three.}

\item{lambda}{parameter four.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{x,q}{vector of quantiles.}
}
\value{
\code{dGMW} gives the density, \code{pGMW} gives the distribution
function, \code{qGMW} gives the quantile function, \code{rGMW}
generates random deviates and \code{hGMW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, random generation and hazard function for the generalized
modified weibull distribution with parameters \code{beta}, \code{theta}, \code{gamma} and \code{lambda}.
}
\details{
The generalized modified weibull with parameters \code{beta}, \code{theta},
\code{gamma} and \code{lambda} has density given by

f(x)= beta*theta*x^(gamma-1)*(gamma+lambda*x)*exp(lambda*x-beta*x^(gamma)*exp(lambda*x))*
(1-exp(beta*x^(gamma)*exp(lambda*x)))^(theta-1)

for x>0.

#'
}
\examples{
## The probability density function
curve(dGMW(x, beta = 2, theta = 0.5, gamma = 2, lambda = 1.5), from = 0, to = 0.8, ylim = c(0, 6), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pGMW(x, beta = 2, theta = 0.5, gamma = 2, lambda = 1.5), from = 0, to = 1.2, col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pGMW(x, beta = 2, theta = 0.5, gamma = 2, lambda = 1.5, lower.tail = FALSE), from = 0, to = 1.2, col = "red", las = 1, ylab = "The Reliability function")
## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qGMW(p, beta = 2, theta = 0.5, gamma = 2, lambda = 0.3), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pGMW(x, beta = 2, theta = 0.5, gamma = 2, lambda = 0.3),  from = 0, add = TRUE, col="red")

## The random function
hist(rGMW(n = 1000, beta = 2, theta = 0.5, gamma = 2,lambda = 0.3), freq = FALSE, xlab = "x", main ="", las = 1)
curve(dGMW(x, beta = 2, theta = 0.5, gamma = 2, lambda = 0.3),  from = 0, add = TRUE, col = "red")

## The Hazard function
curve(hGMW(x, beta = 2, theta = 1.5, gamma = 2, lambda = 0.8), from = 0, to = 1, ylim = c(0, 16), col = "red", ylab = "The Hazard function", las = 1)
}

