% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPW.R
\name{GPW}
\alias{GPW}
\alias{dGPW}
\alias{hGPW}
\alias{pGPW}
\alias{qGPW}
\alias{rGPW}
\title{The Generalized Power Weibull Distribution}
\usage{
dGPW(x, alpha, theta, lambda, log = FALSE)

pGPW(q, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

qGPW(p, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

rGPW(n, alpha, theta, lambda)

hGPW(x, alpha, theta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{theta}{parameter two.}

\item{lambda}{parameter three.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dGPW} gives the density, \code{pGPW} gives the distribution
function, \code{qGPW} gives the quantile function, \code{rGPW}
generates random deviates and \code{hGPW} gives the hazard function.
}
\description{
Density, distribution function, quantile function ,
random generation and hazard function for the generalized power weibull distribution with
parameters \code{alpha}, \code{theta} and \code{lambda}.
}
\details{
The generalized power weibull with parameters \code{alpha}, \code{theta}
and \code{lambda} has density given by

f(x) = alpha*theta*lambda^(-1)*x^(theta-1)*(1+alpha*(x^theta))^(1/lambda-1)*exp(1-(1+alpha*(x^theta))^(1/lambda))

for x > 0.
}
\examples{
## The probability density function
curve(dGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25), from = 0, to = 2.5, ylim = c(0, 3), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25), from = 0, to = 2.5, col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25, lower.tail = FALSE), from = 0, to = 2.5, col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qGPW(p, alpha = 0.5, theta = 0.5, lambda = 0.25), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25), from = 0, add = TRUE, col = "red")

## The random function
hist(rGPW(n = 10000, alpha = 0.5, theta = 0.5, lambda = 0.25), freq = FALSE, xlab = "x", las = 1, main = "")
curve(dGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25),  from = 0, add = TRUE, col = "red")

## The Hazard function
curve(hGPW(x, alpha = 0.5, theta = 0.5, lambda = 0.25), from = 0, to = 6, ylim = c(0, 13), col = "red", las = 1, ylab = "The Hazard function")
}

