% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IW.R
\name{IW}
\alias{IW}
\alias{dIW}
\alias{hIW}
\alias{pIW}
\alias{qIW}
\alias{rIW}
\title{The Inverse Weibull Distribution}
\usage{
dIW(x, alpha, theta, log = FALSE)

pIW(q, alpha, theta, lower.tail = TRUE, log.p = FALSE)

qIW(p, alpha, theta, lower.tail = TRUE, log.p = FALSE)

rIW(n, alpha, theta)

hIW(x, alpha, theta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{##############BUSCAR############}

\item{theta}{##############BUSCAR############}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dIW} gives the density, \code{pIW} gives the distribution
function, \code{qIW} gives the quantile function, and \code{rIW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the inverse weibull distribution with
parameters \code{alpha} and \code{theta}.
}
\details{
The inverse weibull distribution with parameters \code{alpha} and
\code{theta} has density given by

f(x) = alpha*theta*x^(-theta-1)*exp(-alpha*(x^-theta))

for x > 0.
}
\examples{
## Curve
curve(dIW(x,alpha=2,   theta=1), from=0, to=10, ylim=c(0,0.6), col="red",ylab="Density")

## Curve
curve(pIW(x,alpha=2,theta=1), from=0, to=10, ylim=c(0,1), col="blue",ylab="Density")

## Curve
p <- seq(0,0.998, length.out=100)
plot(x=qIW(p,alpha=5,theta=2.5), y=p)

## Curve
hist(rIW(10000,alpha=5,theta=2.5),freq=F,xlab="x", main= "Histogram of rIW")
curve(dIW(x,alpha=5,theta=2.5),  from=0, add=T)

##
curve(hIW(x,alpha=1,   theta=0.6), from=0, to=6, ylim=c(0,1), col="red",ylab="The hazard function")
}

