% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KMW.R
\name{KMW}
\alias{KMW}
\alias{dKMW}
\alias{hKMW}
\alias{pKMW}
\alias{qKMW}
\alias{rKMW}
\title{The Kumaraswamy modified Weibull distribution}
\usage{
dKMW(x, alpha, theta, lambda, a, b, log = FALSE)

pKMW(q, alpha, theta, lambda, a, b, lower.tail = TRUE, log.p = FALSE)

qKMW(p, alpha, theta, lambda, a, b, lower.tail = TRUE, log.p = FALSE)

rKMW(n, alpha, theta, lambda, a, b)

hKMW(x, alpha, theta, lambda, a, b, log = FALSE)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{lambda}{parameter.}

\item{a}{parameter.}

\item{b}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{beta}{parameter.}
}
\value{
\code{dKMW} gives the density, \code{pKMW} gives the distribution
function, \code{qKMW} gives the quantile function, and \code{rKMW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the kumaraswamy modified weibull distribution with
parameters \code{alpha}, \code{theta}, \code{lambda}, \code{a} and \code{b}.
}
\details{
The kumaraswamy modified weibull with parameters \code{alpha}, \code{theta},
\code{lambda}, \code{a} and \code{b} has density given by

f(x)=a*b*alpha*x^(theta-1)*(theta+lambda*x)*exp(lambda*x-alpha*x^(theta)*exp(lambda*x))*
(1-exp(-alpha*x^theta*exp(lambda*x)))^(a-1)*
(1-(1-exp(-aplha*x^theta*exp(lambda*x)))^a)^(b-1)

for x>0.
}
\examples{
## Curve
curve(dKMW(x,alpha=1,theta=0.6,lambda=2,a=2,b=1.2), from=0, to=3, ylim=c(0,2), col="red",ylab="Density")

## Curve
curve(pKMW(x,alpha=1,theta=0.6,lambda=2, a=2,b=1.2), from=0, to=3, col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qKMW(p,alpha=1,theta=0.6,lambda=2, a=2,b=1.2), y=p)

## Curve
hist(rKMW(1000,alpha=1,theta=0.6,lambda=2, a=2,b=1.2),freq=F,xlab="x", main="Histogram of rKMW")
curve(dKMW(x,alpha=1,theta=0.6,lambda=2, a=2,b=1.2),  from=0, add=T)

## Curve
curve(hKMW(x,alpha=8,theta=0.6,lambda=0.01,a=8,b=0.7), from=0, to=3.5, ylim=c(0,10), col="red",ylab="The hazard function")
}

