% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KW.R
\name{KW}
\alias{KW}
\alias{dKW}
\alias{hKW}
\alias{pKW}
\alias{qKW}
\alias{rKW}
\title{The Kumaraswamy Weibull Distribution}
\usage{
dKW(x, alpha, theta, a, b, log = FALSE)

pKW(q, alpha, theta, a, b, lower.tail = TRUE, log.p = FALSE)

qKW(p, alpha, theta, a, b, lower.tail = TRUE, log.p = FALSE)

rKW(n, alpha, theta, a, b)

hKW(x, alpha, theta, a, b)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{theta}{parameter two.}

\item{a}{parameter three.}

\item{b}{parameter four.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dKW} gives the density, \code{pKW} gives the distribution
function, \code{qKW} gives the quantile function, \code{rKW}
generates random deviates and \code{hKW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the kumaraswamy weibull distribution with
parameters \code{alpha}, \code{theta}, \code{a} and \code{b}.
}
\details{
The kumaraswamy weibull with parameters \code{alpha}, \code{theta}, \code{a} and \code{b}
has density given by

f(x)=a*b*alpha*theta*x^(theta-1)*exp(-alpha*x^theta)*(1-exp(-alpha*x^theta))^(a-1)*
(1-(1-exp(-aplha*x^theta))^a)^(b-1)

for x>0.
}
\examples{
## The probability density function
curve(dKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5), from = 0, to = 3, ylim = c(0, 2.5), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5), from = 0, to = 3, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5, lower.tail = FALSE), from = 0, to = 3, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(0,0.99999, length.out=100)
plot(x=qKW(p, alpha = 3, theta = 0.8, a = 2, b = 1.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5), from = 0, add = TRUE, col = "red")

## The random function
hist(rKW(10000, alpha = 3, theta = 0.8, a = 2, b = 1.5), freq = FALSE, xlab = "x", las = 1, ylim = c(0, 2.5), main = "")
curve(dKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5),  from = 0, add = TRUE, col = "red" )

## The Hazard function
curve(hKW(x, alpha = 3, theta = 0.8, a = 2, b = 1.5), from = 0, to = 2.1, ylim = c(0, 4), col = "red", ylab = "The Hazard function", las = 1)
}

