% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LW.R
\name{LW}
\alias{LW}
\alias{dLW}
\alias{pLW}
\alias{qLW}
\alias{rLW}
\title{The Log-Weibull Distribution}
\usage{
dLW(x, a, b, log = FALSE)

pLW(q, a, b, lower.tail = TRUE, log.p = FALSE)

qLW(p, a, b, lower.tail = TRUE, log.p = FALSE)

rLW(n, a, b)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{a}{parameter one.}

\item{b}{parameter two.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dLW} gives the density, \code{pLW} gives the distribution
function, \code{qLW} gives the quantile function, and \code{rLW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the log-weibull distribution with
parameters \code{a} and \code{b}.
}
\details{
The log-weibull distribution with parameters \code{a} and
\code{b} has density given by

f(x)= (1/b)*exp((x-a)/b)*exp(-exp((x-a)/b))

for -Inf < x < Inf.
}
\examples{
## Curve
curve(dLW(x,a = 0,b = 1), from = -20, to = 10, ylim=c(0,0.4), col="red", ylab="Density")

## Curve
curve(pLW(x,a=0,b=1), from=0, to=10, col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qLW(p,a=0,b=3), y=p)

## Curve
hist(rLW(10000,a=0,b=3),freq=F,xlab="x", main= "")
curve(dLW(x,a=0,b=3),  from=-20, to=10, add=T)
}

