% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MW.R
\name{MW}
\alias{MW}
\alias{dMW}
\alias{hMW}
\alias{pMW}
\alias{qMW}
\alias{rMW}
\title{The Modified Weibull Distribution}
\usage{
dMW(x, beta, gamma, lambda, log = FALSE)

pMW(q, beta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

qMW(p, beta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

rMW(n, beta, gamma, lambda)

hMW(x, beta, gamma, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{beta}{shape parameter one.}

\item{gamma}{parameter two.}

\item{lambda}{scale parameter three.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMW} gives the density, \code{pMW} gives the distribution
function, \code{qMW} gives the quantile function, \code{rMW}
generates random deviates and \code{hMW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the modified weibull distribution with
parameters \code{beta}, \code{gamma} and \code{lambda}.
}
\details{
The modified weibull distribution with parameters \code{beta}, \code{gamma}
and \code{lambda} has density given by

f(x) = beta*(gamma+lambda*x)*x^(gamma-1)*exp(lambda*x)*exp(-beta*x^(gamma)*exp(lambda*x))

for x > 0.
}
\examples{
## The probability density function
curve(dMW(x, beta = 2, gamma = 1.5, lambda = 0.2), from=0, to=2, ylim=c(0,1.5), col="red", las=1, ylab="The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pMW(x, beta = 2, gamma = 1.5, lambda = 0.2), from = 0, to = 2, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pMW(x, beta = 2, gamma = 1.5, lambda = 0.2, lower.tail = FALSE), from = 0, to = 2, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x = qMW(p = p, beta = 2, gamma = 1.5, lambda = 0.2), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pMW(x, beta = 2, gamma = 1.5, lambda = 0.2), from = 0, add = TRUE, col = "red")

## The random function
hist(rMW(n = 1000, beta = 2, gamma = 1.5, lambda = 0.2), freq = FALSE, , ylim=c(0,1.5),xlab = "x", las = 1, main = "")
curve(dMW(x, beta = 2, gamma = 1.5, lambda = 0.2), from = 0, , ylim=c(0,1.5), add = T, col = "red")

## The Hazard function
curve(hMW(x, beta = 2, gamma = 1.5, lambda = 0.2), from = 0, to = 1.5, ylim = c(0, 5), col = "red", las = 1, ylab = "The Hazard function")
}

