% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MWEx.R
\name{MWEx}
\alias{MWEx}
\alias{dMWEx}
\alias{hMWEx}
\alias{pMWEx}
\alias{qMWEx}
\alias{rMWEx}
\title{The Modified Weibull Extension Distribution}
\usage{
dMWEx(x, alpha, beta, lambda, log = FALSE)

pMWEx(q, alpha, beta, lambda, lower.tail = TRUE, log.p = FALSE)

qMWEx(p, alpha, beta, lambda, lower.tail = TRUE, log.p = FALSE)

rMWEx(n, alpha, beta, lambda)

hMWEx(x, alpha, beta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{beta}{parameter two.}

\item{lambda}{parameter three.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMWEx} gives the density, \code{pMWEx} gives the distribution
function, \code{qMWEx} gives the quantile function, \code{rMWEx}
generates random deviates and \code{hMWEx} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the modified weibull extension distribution with
parameters \code{alpha}, \code{beta} and \code{lambda}.
}
\details{
The modified weibull extension distribution with parameters \code{alpha}, \code{beta}
and \code{lambda} has density given by

f(x) = (lambda*alpha^((theta-1)/theta)*x^(theta-1)*e^(alpha*x^(theta))*exp(lambda*alpha^(-1/theta))*(1-e^(-alpha*x^theta)))

for x > 0.
}
\examples{
## The probability density function
curve(dMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2), from = 0, to = 2.5, ylim = c(0, 1.5), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2), from = 0, to = 2.5, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2,  lower.tail = FALSE), from = 0, to = 2.5, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x = qMWEx(p, alpha = 1/0.5, beta = 3, lambda = 2), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2), from = 0, add = TRUE, col = "red")

## The random function
hist(rMWEx(n = 10000, alpha = 1/0.5, beta = 3, lambda = 2), freq = FALSE, ylim = c(0, 1.5), xlab = "x", las = 1, main = "")
curve(dMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2),  from = 0, ylim = c(0, 2.5), add = T, col = "red")

## The Hazard function
curve(hMWEx(x, alpha = 1/0.5, beta = 3, lambda = 2), from = 0, to = 1.7, ylim = c(0, 12), col = "red", ylab = "The hazard function", las = 1)
}

