% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PL.R
\name{PL}
\alias{PL}
\alias{dPL}
\alias{hPL}
\alias{pPL}
\alias{qPL}
\alias{rPL}
\title{The Power Lindley Distribution}
\usage{
dPL(x, alpha, beta, log = FALSE)

pPL(q, alpha, beta, lower.tail = TRUE, log.p = FALSE)

qPL(p, alpha, beta, lower.tail = TRUE, log.p = FALSE)

rPL(n, alpha, beta)

hPL(x, alpha, beta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{shape parameter.}

\item{beta}{scale parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dPL} gives the density, \code{pPL} gives the distribution
function, \code{qPL} gives the quantile function, \code{rPL}
generates random deviates and \code{hPL} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the power Lindley distribution with
parameters \code{alpha} and \code{beta}.
}
\details{
The power Lindley distribution with parameters \code{alpha} and
\code{beta} has density given by

f(x) = ((alpha*beta^2)/(beta+1))*(1+x^(alpha))*x^(alpha-1)*exp(-beta*(x^alpha))

for x > 0.
}
\examples{
## The probability density function
curve(dPL(x, alpha = 1, beta = 0.5), from = 0, to = 15, ylim = c(0, 0.25), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pPL(x, alpha = 1, beta = 0.5), from = 0, to = 15,  ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pPL(x, alpha = 1, beta = 0.5, lower.tail = FALSE), from = 0, to = 15,  ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x=qPL(p=p, alpha = 1, beta = 0.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pPL(x, alpha = 1, beta = 0.5), from = 0, add = TRUE, col = "red")

## The random function
hist(rPL(n = 1000, alpha = 1, beta = 0.5), freq = FALSE, , ylim = c(0, 0.25), xlab = "x", las = 1, main = "")
curve(dPL(x, alpha = 1, beta = 0.5),  from = 0, add = T, col = "red", ylim = c(0, 0.25))

## The Hazard function
curve(hPL(x, alpha = 1, beta = 0.5), from = 0, to = 10, ylim = c(0, 0.5), col = "red", las = 1, ylab = "The Hazard function")
}

