% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RW.R
\name{RW}
\alias{RW}
\alias{dRW}
\alias{hRW}
\alias{pRW}
\alias{qRW}
\alias{rRW}
\title{The Reflected Weibull Distribution}
\usage{
dRW(x, alpha, theta, log = FALSE)

pRW(q, alpha, theta, lower.tail = TRUE, log.p = FALSE)

qRW(p, alpha, theta, lower.tail = TRUE, log.p = FALSE)

rRW(n, alpha, theta)

hRW(x, alpha, theta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{theta}{parameter two.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dRW} gives the density, \code{pRW} gives the distribution
function, \code{qRW} gives the quantile function, \code{rRW}
generates random deviates and \code{hRW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the reflected weibull distribution with
parameters \code{alpha} and \code{theta}.
}
\details{
The reflected weibull distribution with parameters \code{alpha} and
\code{theta} has density given by

f(x) = alpha*theta*(-x)^(theta-1)*exp(-alpha*(-x)^alpha)

for - inf < x < 0.
}
\examples{
## The probability density function
curve(dRW(x, alpha = 1, theta = 1), from = -5, to = 0, ylim = c(0, 1), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pRW(x, alpha = 1, theta = 1), from = -5, to = 0, ylim = c(0, 1), col = "red", las = 1, ylab ="The cumulative distribution function")
curve(pRW(x, alpha = 1, theta = 1, lower.tail = FALSE), from = -5, to = 0, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x=qRW(p=p,alpha = 1, theta = 1), y=p, xlab="Quantile", las=1, ylab="Probability")
curve(pRW(x, alpha = 1, theta = 1), from = -5, add = TRUE, col = "red")

## The random function
hist(rRW(n = 10000, alpha = 1, theta = 1), freq = FALSE,xlab = "x", las = 1, main = "")
curve(dRW(x, alpha = 1, theta = 1),  from = -5, to = 0, add = TRUE, col = "red")

## The Hazard function
curve(hRW(x, alpha = 1, theta = 1), from = -5, to = 0, ylim = c(0, 1), col = "red", ylab = "The hazard function", las = 1)
}

