% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SZMW.R
\name{SZMW}
\alias{SZMW}
\alias{dSZMW}
\alias{hSZMW}
\alias{pSZMW}
\alias{qSZMW}
\alias{rSZMW}
\title{The Sarhan and Zaindins Modified Weibull Distribution}
\usage{
dSZMW(x, alpha, beta, gamma, log = FALSE)

pSZMW(q, alpha, beta, gamma, lower.tail = TRUE, log.p = FALSE)

qSZMW(p, alpha, beta, gamma, lower.tail = TRUE, log.p = FALSE)

rSZMW(n, alpha, beta, gamma)

hSZMW(x, alpha, beta, gamma)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{beta}{parameter.}

\item{gamma}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dSZMW} gives the density, \code{pSZMW} gives the distribution
function, \code{qSZMW} gives the quantile function, and \code{rSZMW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for Sarhan and Zaindins modified weibull distribution with
parameters \code{alpha}, \code{beta} and \code{gamma}.
}
\details{
The Sarhan and Zaindins modified weibull with parameters \code{alpha},
\code{beta} and \code{gamma} has density given by

f(x)=(alpha+beta*x*gamma*x^(gamma-1))*exp(-alpha*x-beta*x^gamma)

for x>0.
}
\examples{
## Curve
curve(dSZMW(x,alpha=2,beta=1.5,gamma=0.2), from=0, to=2, ylim=c(0,1.7), col="red",ylab="Density")

## Curve
curve(pSZMW(x,alpha=2,beta=1.5,gamma=0.2), from=0, to=2,  col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qSZMW(p=p,alpha=1.5,beta=3,gamma=9), y=p)

## Curve
hist(rSZMW(1000,alpha=1.5,beta=3,gamma=9),freq=F,xlab="x", main= "Histogram of rSZMW")
curve(dSZMW(x,alpha=1.5,beta=3,gamma=9),  from=0, add=T)

## Curve
curve(hSZMW(x,alpha=1,beta=1.3,gamma=2), from=0, to=3, ylim=c(0,8), col="red",ylab="The hazard function")
}

