% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EW.R
\name{EW}
\alias{EW}
\alias{dEW}
\alias{hEW}
\alias{pEW}
\alias{qEW}
\alias{rEW}
\title{The Exponentiated Weibull Distribution}
\usage{
dEW(x, alpha, theta, lambda, log = FALSE)

pEW(q, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

qEW(p, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

rEW(n, alpha, theta, lambda)

hEW(x, alpha, theta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{scale parameter.}

\item{theta,lambda}{shape parameters.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEW} gives the density, \code{pEW} gives the distribution
function, \code{qEW} gives the quantile function, \code{rEW}
generates random deviates and \code{hEW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation  and hazard function for the exponentiated weibull  distribution with
parameters \code{alpha}, \code{theta} and \code{lambda}.
}
\details{
The Exponentiated Weibull Distribution with parameters \code{alpha},
\code{theta} and \code{lambda} has density given by

f(x)=lambda*alpha*theta*x^(theta-1)*exp(-alpha*(x^theta))*(1-exp(-alpha*(x^theta)))^(lambda-1)

#'
}
\examples{
## The probability density function
curve(dEW(x, alpha = 2, theta = 1.5, lambda = 0.5), from = 0, to = 2, ylim = c(0, 2.5), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pEW(x, alpha = 2, theta = 1.5, lambda = 0.5), from = 0, to = 2,  col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pEW(x, alpha = 2, theta = 1.5, lambda = 0.5, lower.tail = FALSE), from = 0, to = 2,  col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qEW(p, alpha = 2, theta = 1.5, lambda = 0.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pEW(x, alpha = 2, theta = 1.5, lambda = 0.5),  from = 0, add = TRUE, col = "red")

## The random function
hist(rEW(n = 10000, alpha = 2, theta = 1.5, lambda = 0.5), freq = FALSE, xlab = "x", las = 1, main = "")
curve(dEW(x, alpha = 2, theta = 1.5, lambda = 0.5),  from = 0, add = TRUE, col = "red")

## The Hazard function
curve(hEW(x, alpha = 2,theta = 1.5, lambda = 0.5), from = 0, to = 2, ylim = c(0, 7), col = "red", ylab = "The Hazard function")
}

