% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPW.R
\name{GPW}
\alias{GPW}
\alias{dGPW}
\alias{pGPW}
\alias{qGPW}
\alias{rGPW}
\title{The Generalized Power Weibull Distribution}
\usage{
dGPW(x, alpha, theta, lambda, log = FALSE)

pGPW(q, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

qGPW(p, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

rGPW(n, alpha, theta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{theta}{parameter.}

\item{lambda}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dGPW} gives the density, \code{pGPW} gives the distribution
function, \code{qGPW} gives the quantile function, and \code{rGPW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the generalized power weibull distribution with
parameters \code{alpha}, \code{theta} and \code{lambda}.
}
\details{
The generalized power weibull with parameters \code{alpha}, \code{theta}
and \code{lambda} has density given by

f(x) = alpha*theta*lambda^(-1)*x^(theta-1)*(1+alpha*(x^theta))^(1/lambda-1)*exp(1-(1+alpha*(x^theta))^(1/lambda))

for x > 0.
}
\examples{
## Curve
curve(dGPW(x,0.5,0.5,0.25), from=0, to=2.5, ylim=c(0,3), col="red",ylab="Density")

## Curve
curve(pGPW(x,alpha=0.5,theta=0.5,lambda=0.25), from=0, to=2.5,  col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qGPW(p,alpha=0.5,theta=0.5,lambda=0.25), y=p)

## Curve
hist(rGPW(10000,alpha=0.5,theta=0.5,lambda=0.25),freq=F,xlab="x", main= "Histogram of rGEP")
curve(dGPW(x,alpha=0.5,theta=0.5,lambda=0.25),  from=0, add=T)
}

