% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IW.R
\name{IW}
\alias{IW}
\alias{dIW}
\alias{hIW}
\alias{pIW}
\alias{qIW}
\alias{rIW}
\title{The Inverse Weibull Distribution}
\usage{
dIW(x, alpha, theta, log = FALSE)

pIW(q, alpha, theta, lower.tail = TRUE, log.p = FALSE)

qIW(p, alpha, theta, lower.tail = TRUE, log.p = FALSE)

rIW(n, alpha, theta)

hIW(x, alpha, theta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{theta}{parameter two.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dIW} gives the density, \code{pIW} gives the distribution
function, \code{qIW} gives the quantile function, \code{rIW}
generates random deviates and \code{hIW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the inverse weibull distribution with
parameters \code{alpha} and \code{theta}.
}
\details{
The inverse weibull distribution with parameters \code{alpha} and
\code{theta} has density given by

f(x) = alpha*theta*x^(-theta-1)*exp(-alpha*(x^-theta)),

for x > 0.
}
\examples{
## The probability density function
curve(dIW(x, alpha = 5, theta = 2.5), from = 0, to = 10, ylim = c(0, 0.55), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pIW(x, alpha = 5, theta = 2.5), from = 0, to = 10, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pIW(x, alpha = 5, theta = 2.5, lower.tail = FALSE), from = 0, to = 10, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x = qIW(p, alpha = 5, theta = 2.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pIW(x, alpha = 5, theta = 2.5),  from = 0, add = TRUE, col = "red")

## The random function
hist(rIW(n = 1000, alpha = 5, theta = 2.5), freq = FALSE, xlab = "x", ylim = c(0, 0.55), las = 1, main = "")
curve(dIW(x, alpha = 5, theta = 2.5),  from = 0, to = 10, add = TRUE, ylim = c(0, 0.55), col = "red")

## The Hazard function
curve(hIW(x, alpha = 5, theta = 2.5), from = 0, to = 15, ylim = c(0, 1), col = "red", las = 1, ylab = "The Hazard function")
}

