% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KW.R
\name{KW}
\alias{KW}
\alias{dKW}
\alias{hKW}
\alias{pKW}
\alias{qKW}
\alias{rKW}
\title{The Kumaraswamy Weibull Distribution}
\usage{
dKW(x, alpha, theta, a, b, log = FALSE)

pKW(q, alpha, theta, a, b, lower.tail = TRUE, log.p = FALSE)

qKW(p, alpha, theta, a, b, lower.tail = TRUE, log.p = FALSE)

rKW(n, alpha, theta, a, b)

hKW(x, alpha, theta, a, b)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{a}{parameter.}

\item{b}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{beta}{parameter.}
}
\value{
\code{dKW} gives the density, \code{pKW} gives the distribution
function, \code{qKW} gives the quantile function, and \code{rKW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the kumaraswamy weibull distribution with
parameters \code{alpha}, \code{beta}, \code{a} and \code{b}.
}
\details{
The kumaraswamy weibull with parameters \code{alpha}, \code{beta}, \code{a} and \code{b}
has density given by

f(x)=a*b*alpha*theta*x^(theta-1)*exp(-alpha*x^theta)*(1-exp(-alpha*x^theta))^(a-1)*
(1-(1-exp(-aplha*x^theta))^a)^(b-1)

for x>0.
}
\examples{
## Curve
curve(dKW(x,alpha=3,beta=0.8,a=2,b=1.5), from=0, to=3, ylim=c(0,2.5), col="red",ylab="Density")

## Curve
curve(dKW(x,alpha=3,theta=0.8,a=2,b=1.5), from=0, to=3, ylim=c(0,2.5), col="red",ylab="Density")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qKW(p,alpha=3,theta=0.8,a=2,b=1.5), y=p)

## Curve
hist(rKW(10000,alpha=1.2,theta=1.5,a=2,b=1.2),freq=F,xlab="x", main= "Histogram of rKW")
curve(dKW(x,alpha=1.2,theta=1.5,a=2,b=1.2),  from=0, add=T)

## Curve
curve(hKW(x,alpha=3,theta=0.5,a=3,b=3), from=0, to=2.1, ylim=c(0,10), col="red",ylab="The hazard function")
}

