% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MW.R
\name{MW}
\alias{MW}
\alias{dMW}
\alias{hMW}
\alias{pMW}
\alias{qMW}
\alias{rMW}
\title{The Modified Weibull Distribution}
\usage{
dMW(x, beta, gamma, lambda, log = FALSE)

pMW(q, beta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

qMW(p, beta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

rMW(n, beta, gamma, lambda)

hMW(x, beta, gamma, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{beta}{shape parameter.}

\item{gamma}{############## NO DICE NADA SOBRE ESTE PARÁMETRO}

\item{lambda}{scale parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMW} gives the density, \code{pMW} gives the distribution
function, \code{qMW} gives the quantile function, and \code{rMW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the modified weibull distribution with
parameters \code{beta}, \code{gamma} and \code{lambda}.
}
\details{
The modified weibull distribution with parameters \code{beta}, \code{gamma}
and \code{lambda} has density given by

f(x) = beta*(gamma+lambda*x)*x^(gamma-1)*exp(lambda*x)*exp(-beta*x^(gamma)*exp(lambda*x))

for x > 0.
}
\examples{
## Curve
curve(dMW(x,2,1.5,0.2), from=0, to=2, ylim=c(0,2.2), col="red",ylab="Density")

## Curve
curve(pMW(x, beta=2,gamma=1.5,lambda=0.2), from=0, to=2, col="blue")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qMW(p=p,beta=2,gamma=1.5,lambda=0.2), y=p)

## Curve
hist(rMW(1000,beta=2,gamma=1.5,lambda=0.2),freq=F,xlab="x", main= "Histogram of rMW")
curve(dMW(x,beta=2,gamma=1.5,lambda=0.2),  from=0, add=T)

##
curve(hMW(x,beta=2,gamma=1.5,lambda=0.2), from=0, to=1.5, ylim=c(0,5), col="red",ylab="Density")
}

