% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MWEx.R
\name{MWEx}
\alias{MWEx}
\alias{dMWEx}
\alias{pMWEx}
\alias{qMWEx}
\alias{rMWEx}
\title{The Modified Weibull Extension Distribution}
\usage{
dMWEx(x, alpha, beta, lambda, log = FALSE)

pMWEx(q, alpha, beta, lambda, lower.tail = TRUE, log.p = FALSE)

qMWEx(p, alpha, beta, lambda, lower.tail = TRUE, log.p = FALSE)

rMWEx(n, alpha, beta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{beta}{parameter.}

\item{lambda}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMWEx} gives the density, \code{pMWEx} gives the distribution
function, \code{qMWEx} gives the quantile function, and \code{rMWEx}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the modified weibull extension distribution with
parameters \code{alpha}, \code{beta} and \code{lambda}.
}
\details{
The modified weibull extension distribution with parameters \code{alpha}, \code{beta}
and \code{lambda} has density given by

f(x) =

for x > 0.
}
\examples{
## Curve
curve(dMWEx(x,alpha=1/0.5,beta=3,lambda=2),from=0, to=2.5, ylim=c(0,2.5), col="red", ylab="Density")

## Curve
curve(pMWEx(x,alpha=1/0.5,beta=3,lambda=2),from=0, to=2.5, col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qMWEx(p,alpha=1/0.7,beta=2,lambda=0.2), y=p)

## Curve
hist(rMWEx(10000,alpha=1/0.7,beta=2,lambda=0.2),freq=F,xlab="x", main= "Histogram of rGEP")
curve(dMWEx(x,alpha=1/0.7,beta=2,lambda=0.2),  from=0, add=T)
}

