% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NMW.R
\name{NMW}
\alias{NMW}
\alias{dNMW}
\alias{hNMW}
\alias{pNMW}
\alias{qNMW}
\alias{rNMW}
\title{The Almaki and Yuans modified Weibull distribution}
\usage{
dNMW(x, alpha, beta, theta, gamma, lambda, log = FALSE)

pNMW(q, alpha, beta, theta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

qNMW(p, alpha, beta, theta, gamma, lambda, lower.tail = TRUE, log.p = FALSE)

rNMW(n, alpha, beta, theta, gamma, lambda)

hNMW(x, alpha, beta, theta, gamma, lambda, log = FALSE)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{beta}{parameter.}

\item{theta}{parameter.}

\item{gamma}{parameter.}

\item{lambda}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dNMW} gives the density, \code{pNMW} gives the distribution
function, \code{qNMW} gives the quantile function, and \code{rNMW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the Almaki and Yuans modified weibull distribution with
parameters \code{alpha}, \code{beta}, \code{theta}, \code{gamma} and \code{lambda}.
}
\details{
The Almaki and Yuans modified weibull with parameters \code{alpha},
\code{beta}, \code{theta}, \code{gamma} and \code{lambda} has density given by

f(x)=(alpha*theta*x^(theta-1)+beta*(gamma+ lambda*x)*(x^(gamma-1)*exp(lambda*x)))*(exp((-alpha*x^theta)-(beta*x^gamma*exp(lambda*x)))

for x>0.
}
\examples{
## Curve
curve(dNMW(x,alpha=1.15,beta=0.15,theta=0.75,gamma=5,lambda=2), from=0, to=1.4, ylim=c(0,3), col="red",ylab="Density")

## Curve
curve(pNMW(x,alpha=1.15,beta=0.15,theta=0.75,gamma=5,lambda=2), from=0, to=1.4,  col="red")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qNMW(p,alpha=1.15,beta=0.15,theta=0.75,gamma=5,lambda=2), y=p)

## Curve
hhist(rNMW(1000,alpha=1.15,beta=0.15,theta=0.75,gamma=5,lambda=2),freq=F,xlab="x", main="")
curve(dNMW(x,alpha=1.15,beta=0.15,theta=0.75,gamma=5,lambda=2),  from=0, add=T)

## Curve
curve(hNMW(x,alpha=1.2,beta=1.5,theta=3,gamma=0.5,lambda=0.75), from=0, to=1.5, ylim=c(0,8), col="red",ylab="The hazard function")
}

