% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OW.R
\name{OW}
\alias{OW}
\alias{dOW}
\alias{pOW}
\alias{qOW}
\alias{rOW}
\title{The Odd Weibull Distribution}
\usage{
dOW(x, alpha, theta, lambda, log = FALSE)

pOW(q, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

qOW(p, alpha, theta, lambda, lower.tail = TRUE, log.p = FALSE)

rOW(n, alpha, theta, lambda)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter.}

\item{theta}{parameter.}

\item{lambda}{parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dOW} gives the density, \code{pOW} gives the distribution
function, \code{qOW} gives the quantile function, and \code{rOW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the odd weibull distribution with
parameters \code{alpha}, \code{theta} and \code{lambda}.
}
\details{
The generalized power weibull with parameters \code{alpha}, \code{theta}
and \code{lambda} has density given by

f(x) = alpha*theta*lambda*x^(theta-1)*exp(alpha*(x^theta))*(exp(alpha*(x^theta))*-1)^(lambda-1)*(1+(exp(alpha*(x^theta))-1)^lambda)^-2
### revisar
for x > 0.
}
\examples{
## Curve
curve(dOW(x,2,3,0.2), from=0, to=4, ylim=c(0,2), col="blue",ylab="Density")

## Curve
curve(pOW(x, alpha=2,theta=3,lambda=0.2), from=0, to=4, col="blue")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qOW(p,alpha=2,theta=3,lambda=0.2), y=p)

## Curve
hist(rOW(10000,alpha=2,theta=3,lambda=0.2),freq=F,xlab="x", main= "Histogram of rOW")
curve(dOW(x,alpha=2,theta=3,lambda=0.2),  from=0, add=T)
}

