% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PL.R
\name{PL}
\alias{PL}
\alias{dPL}
\alias{hPL}
\alias{pPL}
\alias{qPL}
\alias{rPL}
\title{The Power Lindley Distribution}
\usage{
dPL(x, alpha, beta, log = FALSE)

pPL(q, alpha, beta, lower.tail = TRUE, log.p = FALSE)

qPL(p, alpha, beta, lower.tail = TRUE, log.p = FALSE)

rPL(n, alpha, beta)

hPL(x, alpha, beta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{shape parameter.}

\item{beta}{scale parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dPL} gives the density, \code{pPL} gives the distribution
function, \code{qPL} gives the quantile function, and \code{rPL}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the power Lindley distribution with
parameters \code{alpha} and \code{beta}.
}
\details{
The power Lindley distribution with parameters \code{alpha} and
\code{beta} has density given by

f(x) = (((alpha*beta)^2)/(beta+1))*(1+x^(alpha))*x^(alpha-1)*exp(-beta*(x^alpha))

for x > 0.
}
\examples{
## Curve
curve(dPL(x,1,0.5), from=0, to=15, ylim=c(0,0.6), col="red",ylab="Density")

## Curve
curve(pPL(x,alpha=1,beta=0.5), from=0, to=15,  ylim=c(0,1), col="blue")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qPL(p=p,alpha=0.8,beta=0.3), y=p)

## Curve
hist(rPL(1000,alpha=2,beta=0.5),freq=F,xlab="x", main= "Histogram of rGEP")
curve(dPL(x,alpha=2,beta=0.5),  from=0, add=T)

##
curve(hPL(x,alpha=0.2,beta=1), from=0, to=10, ylim=c(0,1.5), col="red",ylab="The hazard function")
}

