% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RW.R
\name{RW}
\alias{RW}
\alias{dRW}
\alias{pRW}
\alias{qRW}
\alias{rRW}
\title{The Reflected Weibull Distribution}
\usage{
dRW(x, alpha, theta, log = FALSE)

pRW(q, alpha, theta, lower.tail = TRUE, log.p = FALSE)

qRW(p, alpha, theta, lower.tail = TRUE, log.p = FALSE)

rRW(n, alpha, theta)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{##############BUSCAR############}

\item{theta}{##############BUSCAR############}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dRW} gives the density, \code{pRW} gives the distribution
function, \code{qRW} gives the quantile function, and \code{rRW}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the reflected weibull distribution with
parameters \code{alpha} and \code{theta}.
}
\details{
The reflected weibull distribution with parameters \code{alpha} and
\code{theta} has density given by

f(x) = alpha*theta*(-x)^(theta-1)*exp(-alpha*(-x)^alpha)

for - inf < x < 0.  ### También x < 0.
}
\examples{
## Curve
curve(dRW(x,1,1), from=-5, to=0, ylim=c(0,1), col="red",ylab="Density")

## Curve
curve(pRW(x,alpha=1,theta=1), from=-5, to=0, ylim=c(0,1), col="blue",ylab="Density")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qRW(p=p,alpha=0.5,theta=2), y=p)

## Curve
hist(rRW(10000,alpha=0.5,theta=2), freq=F,xlab="x", main= "Histogram of rRW")
curve(dRW(x,alpha=0.5,theta=2),  from=-5, to=0, add=T)
}

