% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SZMW.R
\name{SZMW}
\alias{SZMW}
\alias{dSZMW}
\alias{hSZMW}
\alias{pSZMW}
\alias{qSZMW}
\alias{rSZMW}
\title{The Sarhan and Zaindins Modified Weibull Distribution}
\usage{
dSZMW(x, alpha, beta, gamma, log = FALSE)

pSZMW(q, alpha, beta, gamma, lower.tail = TRUE, log.p = FALSE)

qSZMW(p, alpha, beta, gamma, lower.tail = TRUE, log.p = FALSE)

rSZMW(n, alpha, beta, gamma)

hSZMW(x, alpha, beta, gamma)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{parameter one.}

\item{beta}{parameter two.}

\item{gamma}{parameter three.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dSZMW} gives the density, \code{pSZMW} gives the distribution
function, \code{qSZMW} gives the quantile function, \code{rSZMW}
generates random deviates and \code{hSZMW} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for Sarhan and Zaindins modified weibull distribution with
parameters \code{alpha}, \code{beta} and \code{gamma}.
}
\details{
The Sarhan and Zaindins modified weibull with parameters \code{alpha},
\code{beta} and \code{gamma} has density given by

f(x)=(alpha+beta*gamma*x^(gamma-1))*exp(-alpha*x-beta*x^gamma)

for x>0.
}
\examples{
## The probability density function
curve(dSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2), from = 0, to = 2, ylim = c(0, 1.7), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2), from = 0, to = 2, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2, lower.tail = FALSE), from = 0, to = 2, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qSZMW(p = p, alpha = 2, beta = 1.5, gamma = 0.2), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2), from = 0, add = TRUE, col = "red")

## The random function
hist(rSZMW(n = 1000, alpha = 2, beta = 1.5, gamma = 0.2), freq = FALSE, xlab = "x", las = 1, main = "")
curve(dSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2),  from = 0, add = TRUE, col = "red")

## The Hazard function
curve(hSZMW(x, alpha = 2, beta = 1.5, gamma = 0.2), from = 0, to = 3, ylim = c(0, 8), col = "red", ylab = "The hazard function", las = 1)
}

