% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WG.R
\name{WG}
\alias{WG}
\alias{dWG}
\alias{hWG}
\alias{pWG}
\alias{qWG}
\alias{rWG}
\title{The Weibull Geometric Distribution}
\usage{
dWG(x, alpha, beta, pi, log = FALSE)

pWG(q, alpha, beta, pi, lower.tail = TRUE, log.p = FALSE)

qWG(p, alpha, beta, pi, lower.tail = TRUE, log.p = FALSE)

rWG(n, alpha, beta, pi)

hWG(x, alpha, beta, pi)
}
\arguments{
\item{x,q}{vector of quantiles.}

\item{alpha}{shape parameter.}

\item{beta}{scale parameter.}

\item{pi}{parameter of geometric random variable.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dWG} gives the density, \code{pWG} gives the distribution
function, \code{qWG} gives the quantile function, and \code{rWG}
generates random deviates.
}
\description{
Density, distribution function, quantile function and
random generation for the weibull geometric distribution with
parameters \code{alpha}, \code{beta} and \code{pi}.
}
\details{
The weibull geometric distribution with parameters \code{alpha},
\code{beta} and \code{pi} has density given by

f(x) = ((alpha*beta)^alpha*(1-pi)*x^(alpha-1)*exp(-(beta*x)^alpha))/(1- pi*exp(-(beta*x)^alpha))^2

for x > 0.
}
\examples{
## Curve
curve(dWG(x,0.5,0.2, 0.95), from=0, to=5, ylim=c(0,1.5), col="red",ylab="Density")

## Curve
curve(pWG(x,0.5,0.2, 0.95), from=0, to=5, col="blue")

## Curve
p <- seq(0,0.99999, length.out=100)
plot(x=qWG(p=p,alpha=0.5,beta=0.2, pi=0.95), y=p)

## Curve
hist(rWG(1000,alpha=1.5,beta=0.2, pi=0.95),freq=F,xlab="x", main= "Histogram of rWG")
curve(dWG(x,alpha=1.5,beta=0.2, pi=0.95),  from=0, add=T)

##
curve(hWG(x,alpha=2,beta=0.25, pi=0.15), from=0, to=15, ylim=c(0,2.2), col="red",ylab="The hazard function")
}

